Add left position for Ff transitions
Firefox's transition support is detected, but not its transform
support; result being that the transition is based on CSS position
rather than translation.

Obviously that needs to be corrected, but my immediate issue is
that because no initial left position is set, there's no transition
when `.next()` is called before manual swiping and consequently no
transitionend event. That's a problem when using pagination like in
the first gallery demo.